The building commodity costs map now uses int64_t for its values

+++ src/economy/commodity.h
@@ -181,10 +181,10 @@ public:
 
 	std::pair<std::string, const commodity_unit *> string_to_number_string_and_unit(const std::string &str) const;
 	decimillesimal_int string_to_fractional_value(const std::string &str) const;
-	int string_to_value(const std::string &str) const;
+	int64_t string_to_value(const std::string &str) const;
 	std::pair<std::variant<int, dice>, const commodity_unit *> string_to_value_variant_with_unit(const std::string &str) const;
 
-	std::string value_to_string(const int value) const;
+	std::string value_to_string(const int64_t value) const;
 	std::string value_to_string(const centesimal_int &value) const;
 	Q_INVOKABLE QString value_to_qstring(const int value) const;
